pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rice-blast-env"
version = "1.0.0"
description = "OpenEnv-compliant RL environment for rice blast disease early detection"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic>=2.7.1",
    "numpy>=1.26.4",
    "pyyaml>=6.0.1",
    "openai>=1.30.1",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.2.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
server = "server.app:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hypothesis]
max_examples = 100
